0.16.12 - Moving Display List to component
Didn't figure it out totally but got the code moved to a separate file and rendering the list and adding items. The issue is with the edit and delete buttons. The delete button doesn't work at all and the edit button only shows undefined in the box. I understand why it says this but I couldnt get a solution for this.