0.18 - SImple Slider 2
I've completely rebuilt and approached the slider from a different angle. First by ensuring that the images fit perfectly inside the container and then making sure that the slider changed height according with the size of the window, keeping the images at the same dimensions(important for the slide functionality). I have not yet added the controls but these where the 2 biggest problems in the first slider so that's an improvement.