1.10.17 - Started work on custom map tooltip
Instead of using the default tooltip from leaflet, i am using my own custom tooltip in the form of a full panel of information. This should be much more useful than the default option. I have created the panel and relevant event handling so the next step is to populate the panel with information and add styling