1.10.17.1 - Added more information to panel
The reuse of code is something im not sure how to get around. I have different information that need to be styled slightly differently. I know that i could use a map to output the items but i dont think that would allow me the same level of control over what I am doing right now. This is an issue for all of these slides when outputting API data. Im not sure if my approach is the best but if i can think of something i will come back and make any improvements.