1.11.4 - RPM - Added hide project button
Added a button that lets you hide tasks for whatever reason. Probably more useful on mobile but I thought it would be a nice feature. It currently resets on page load so is limited in actual usefulness but it's a nice tool in allowing user control of the page. I still need to style it.