1.13.14 - Started again
I actually solved the problem with the previous variation( a misplaced JSON.stringify) but decided to start a fresh. I was getting a bit irritated at trying to refactor the code into what I felt was the correct way forward. Starting a fresh removes all of that strain and lets me focus more easily on the drag and drop library