1.13.15 - Caught Up
I have implemented what I think to be a better version of the project(finally). I am using a single item array with a board object that changes as you drag and drop onto each board(instead of the separate arrays as I using previously) This approach solves some of the issues i was having and is a lot smaller in code size. I still need to clean up the code as it's all in one file but I am back to where I was in terms of project functionality with everything dragging and dropping as expected.