1.13.28 - Basic Drag and drop with react-beautiful-dnd
So react beautiful dnd doesn't work with my current react version or react strict mode so there's been a few hoops to jump through, but I did manage to create a simple list that can be dragged around and re-arranged. I need to mess around with this to really understand what's going on before I move on.