1.13.40 - Figured it out
I did figure out a better way of doing things as I was over complicating things a bit. I was doing an Object.entries twice for no reason to identify the unnassigned column. I already did this in the user input form so i just passed the previous value into a map which removed the need for the double Object.entries. Long story short i now have a way of adding to the board using user input.